Update the layout structure
